API support and fetch data -2
