+++ SPA/views/pages/Interviewees.js
@@ -1,22 +1,4 @@
-let getIntervieweesList = async () => {
-    const options = {
-       method: 'GET',
-       headers: {
-           'Content-Type': 'application/json'
-       }
-   };
-   try {
-    const response = await fetch(`http://localhost:3000/interviewees/`, options)
-    const json = await response.json();
-    console.log("p")
-    console.log(json)
-    return json
-} catch (err) {
- console.log("p")
-    console.log('Error getting documents', err)
-}
-}
-
+import getIntervieweesList        from '../../services/getIntervieweesList.js'
 
 
 window.deleteInterviewee = async (id) => {
