+++ SPA/views/pages/Interviewers.js
@@ -1,23 +1,5 @@
 import Redirect        from '../../services/Redirect.js'
-let getInterviewersList = async () => {
-    const options = {
-       method: 'GET',
-       headers: {
-           'Content-Type': 'application/json'
-       }
-   };
-   try {
-       const response = await fetch(`http://localhost:3000/interviewers/`, options)
-       const json = await response.json();
-       console.log("p")
-       console.log(json)
-       return json
-   } catch (err) {
-    console.log("p")
-       console.log('Error getting documents', err)
-   }
-  }
-
+import getInterviewersList        from '../../services/getInterviewersList.js'
 
 
   window.deleteInterviewer = async (id) => {
